:heavy_plus_sign: Add Surveys class
