+++ redcap/methods/__init__.py
@@ -1,0 +1,11 @@
+"""Make the method modules available upon import"""
+
+import redcap.methods.field_names
+import redcap.methods.files
+import redcap.methods.instruments
+import redcap.methods.metadata
+import redcap.methods.project_info
+import redcap.methods.records
+import redcap.methods.reports
+import redcap.methods.surveys
+import redcap.methods.users
